+++ app/actions/HomeActions.js
@@ -7,6 +7,9 @@ class HomeActions {
       'getDeptListFail',
       'getCoursesSuccess',
       'getCoursesFail',
+      'searchSuccess',
+      'searchFail',
+      'updateDeptSelected',
       'updateDepts',
       'updateColumn',
       'filterTime'
@@ -24,7 +27,6 @@ class HomeActions {
   }
 
   getCourses(dept) {
-    dept = dept.split('=')[1];
     $.ajax({ url: '/api/course/' + dept })
       .done(data => {
         this.actions.getCoursesSuccess(data);
@@ -35,6 +37,35 @@ class HomeActions {
       });
   }
 
+  search(q, dept) {
+    var obj = {};
+    var query = {};
+    var bool = {};
+    if(dept)
+      bool.must = {match: {dept_no: dept}};
+    bool.should = [];
+    bool.should.push({match: {dept_no: q}});
+    bool.should.push({match: {course_no: q}});
+    bool.should.push({match: {name: q}});
+    bool.should.push({match: {teacher: q}});
+    bool.should.push({match: {required: q}});
+    bool.should.push({match: {memo: q}});
+    query.bool = bool;
+    obj.query = query;
+    console.log(JSON.stringify(obj));
+    $.ajax({
+      url: '/api/search',
+      method: "POST",
+      data: JSON.stringify(obj),
+      contentType: 'application/json; charset=utf-8',
+    }).done(data => {
+        this.actions.searchSuccess(data);
+      })
+      .fail(jqXhr => {
+        console.log(jqXhr);
+        this.actions.searchFail("載入失敗，請重試");
+      });
+  }
 }
 
 export default alt.createActions(HomeActions);